Add and remove cars

+++ UCTS.Manager.BL/RideStatistics.cs
@@ -14,7 +14,7 @@ namespace UCTS.Manager.BL
         public int RideNumber { get; set; }
         public int NumberOfPassengers { get; set; }
         public double TravelDistance { get; set; }
-        public decimal CostPerKm { get; set; }
+        public double CostPerKm { get; set; }
         ///Total income: sum of(passengers num * travel distance * cost per km) for all travels
         public decimal TotalIncome { get; set; }
     }
